Fix broken link - links are case sensitive for this server!
As mentioned in the title, it looks like links for this server are case sensitive.  I'd suggest trying to shake the habit of using capitalization OR spaces in file names / paths as they're often quite confusing.  It's much easier to just never use a space (instead using an underscore _) and never use a capital.  That said, also good to know how it works because sometimes we don't have control over naming.

Another tip; I tested this by clicking the link & see it was broken.  I then read the URL and tried to cross reference that w/ your repo.  I saw this as the broken link in my browser "https://notmetaseth.github.io/MetaSeth.github.io/the_autodidactic_programmer.html".  It said "404 page not found", so I know it failed looking for that file.  "the_autodidactic_programmer.html" was the file name in the path, so when I looked at your repo & saw it w/ capitals, I knew that was the issue.